maybe have simplex working.
not sure since i have only one test and I have not verified it.

put aside trying to parameterize Graph over its associated VertexTable
(in terms of whether e.g. it carries a hash map or not), and just attached
a HashMap to Graph.

Even so, the simplex code is not simple at all, because I had to
divide the single pass into two (maybe three) separate ones: the first
to initialize each vertex's name before moving the vertex vector into
the area arena, the second to initialize the hash map based on each
vertex's name, and the third to recompute the same indices again but
this time for us in computing edges.

The latter issue is a place where a Reap like structure might be nice:
something where I have mutabe ownership when I first get the object
from the Reap, but then I can downgrade it to shared access; a reap
structure would hopeully let me combine the three passes above back
into a single pass (assuming I can make the lifetimes work out, but it
seems reasonable in this context).